requirements.txt
flask
google-generativeai
pillow
sqlite3